TdbHdf5Connection: Fixed a file descriptor leak, an assertion and a comment.
The database was not closed after the file had been unlinked (or deleted) from
the filesystem. This caused the file to be kept open, but never closed properly.
Apparently the correct way to close it is not to only unreference it from the
m_tableFiles map, but also close it, i.e. as done by closeTableFile(). While
testing this solution, I found that it is most likely that closeTableFile() had
never been called at all, because the assertion its code started with was plain
incorrect and would most likely have been triggered.